[feat] post-hooks to support external chezmoi repo for highly shared files
AWS is a particular motivator here, since: this is a highly shared config file that could morph over time as adoption increases and will never have personal accounts. Adding a post-hook to the personal config allows an external chezmoi to act in its own namespace and act as part of existing personal workflows.